update
오랜만에 프로그래밍을 잡음 ㅎㅎㅎㅎ

+++ CodeUp/10XX/1090 - 수 나열하기2/source.cpp
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-	int a,r,n;
+	long long int a,r,n;
 	scanf("%d %d %d", &a, &r, &n);
-	for(int i=1; i<n; i++, a*=r);
-	printf("%d", a);
+	for(int i=1; i<n; i++) a*=r;
+	printf("%lld", a);
 	return 0;
 }
